charted through firing process. All necessary functions are present, no extraneous ones remain as far as I can see.
git-svn-id: https://elysia.svn.sourceforge.net/svnroot/elysia/trunk@839 e0ac87d7-b42b-0410-9034-8248a05cb448

+++ brain/test.cpp
@@ -17,7 +17,7 @@ void test1() {
     dummy = myProteinEnvironment;
 
 	Brain *brain= new Brain(dummy);
-	for(int i=0;i<100;i++){
+	for(float i=0.0f;i<100;i++){
 		Vector3f v;
 		v.x = i;
 		v.y = i;
